Create CHANGELOG.md
A Changelog file gives the developer a place to describe relevant changes to code/algorithms/strategy/etc. It can incorporate the reasoning behind the changes without burdening the top-level README with lots of detail. It also lets others catch up to the current state of the project if they've been away for a while.