Add control to ignore delay samples by sequence number for plotting
On intialisation new reflectors start with a high artificial baseline
that will cause delay plot auto scaling to scale the y-axis to
too high values. To avoid this allow to require a minimum sequence
number for the delay samples and only plot samples with a sequence number
>= the required minimum.
Set to zero to include all samples, setting this to one should already remove
the biggest values. For ease of implementation these samples are not plotted at
all and not just excluded from autoscaling.

Signed-off-by: Sebastian Moeller <moeller0@gmx.de>