More robust reflector rotation process handling
On each pinger process termination the corresponding monitor process is automatically terminated and the baselines and delta ewmas are written out. The automatic termination relies upon the while read loop ending because the fifo to be read was deleted. 

But sometimes on reflector rotation a new monitor process would get created before the old monitor process had fully terminated so that the baselines and delta ewmas were only half written. Worse still, the automatic termination could break depending on timing because a new fifo gets created, resulting in redundant monitoring processes hanging around. 

So now we simply wait for the monitor processes to fully end upon pinger process termination. And this way when a new monitor process is created the old one should have fully terminated.