Robustify the parser
Sometimes log files start with a stray DATA or LOAD record with the matching
DATA_HEADER or LOAD_HEADER following later. We used to refrain parsing such files at all,
but now we simply SKIP over the records preceeding the respective header record.
This still fails when there is no header, but should deal with the way current logfiles are written.
This commit also adds a specified reflector subset in the filename.

Signed-off-by: Sebastian Moeller <moeller0@gmx.de>