Only determine ewma delta on low load and add reflector comparison checks and new data field
This commit includes further reflector health checks. Now a reflector is randomly rotated every 60 minutes. And every minute a reflector is rotated if its baseline exceeds the minimum baseline by a set threshold or if its delta ewma exceeds a compensated limit so that it is too close to triggering on low load.