Fix plotting of PDFs.
In addition ot CDFs this allows to also plot PDF/histograms of the delay distributions.
However at least for the flent delay source the time quantisation requires a bin-size
of 200µs, and the resulting distributions are really small, so at least for low load
these plots are deswkected in te default set.

Signed-off-by: Sebastian Moeller <moeller0@gmx.de>