Instrument kill_pinger and increase verbosity across the board
This adds a number of additional debug messages and rigs up
the kill commands in kill_pinger() to use cmd_wrapper() to
log both success and failure. This also fixes an issue when terminating
cake-autorate_launcher.sh with CTRL-C; this did not trigger
the trapped_INT logic for some reason or the other.
THe "fix" is to set SILENT_TERM for all of INT, TERM, EXIT
In addition this tries to make cake-autorate_launcher.sh more verbose in what
it starts and what it kills and and logs these to the system log.)

Signed-off-by: Sebastian Moeller <moeller0@gmx.de>