Make break from outstanding IPC reads conditional
It is desirable to break out of reading outstanding IPC commands on receipt of a request to change state to START or STOP in order to start or stop the pingers as soon as possible. But it is undesirable to break out of reading outstanding IP commands on receipt of a request to change state to RUNNING or PAUSED because those commands may legitimately cancel each other out on a stall and recovery therefrom and also because outstanding reflector timestamps should be read in and processed before switching to the RUNNING state to avoid inappropriately tagging any reflectors as misbehaving.