Cleanup and bug in parser for iputils-ping
Reduce nesting and slight fix of inconsistent tabs/space which
need some work to be fully "fixed."

While cleaning up, a bug was found where the iputils-ping parser
was checking if ts == cksum but not skipping the loop if it were
the case. This was addressed along with the general cleanup.

Signed-off-by: rany <ranygh@riseup.net>