Don't use ${SHELL} and just call the script directly
There is no clean way to get the current shell. They all would
require you to check /proc/$$/comamnd and I don't want to parse it.
It's more complicated with the tools that I have in the setup.sh script
and the easier approaches would not be proper (replacing \0 with \n, calling
head, etc.).

Signed-off-by: rany <ranygh@riseup.net>