Fix install instructions for OpenWRT and AsusWRT
`opkg update` needs to run, if you're following these instructions
you probably will be confused by the errors it would return if the
package list was never fetched.

Also drop `iputils-ping` as it is terrible and shouldn't be supported
by cake-autorate anyway. There is no reason someone would opt for that
over fping or tsping.

Fixes #306

Signed-off-by: rany <ranygh@riseup.net>